+++ common/unit.h
@@ -144,7 +144,7 @@ struct unit {
 
   int upkeep[O_LAST]; // unit upkeep with regards to the homecity
 
-  int moves_left;
+  float moves_left;
   int hp;
   int veteran;
   int fuel;
